refactor: cached.rs / refresh.rs を削除し main.rs にインライン化
bin 層の薄いモジュールを composition root（main.rs）に統合する。
DaemonCacheAdapter と daemon refresh ハンドラを main.rs に直接記述。
