+++ src/main.rs
@@ -1,6 +1,4 @@
-mod cached;
 mod contexts;
-mod refresh;
 
 use clap::{CommandFactory, Parser, Subcommand};
 use contexts::configuration::application::config_service::ConfigService;
@@ -8,13 +6,16 @@ use contexts::configuration::infrastructure::toml_loader::TomlConfigRepository;
 use contexts::merge_readiness::application::{
     OutputToken,
     errors::ErrorToken,
-    prompt::{ExecutionMode, RepoIdPort},
+    prompt::{ExecutionMode, PromptEffect, RepoIdPort},
 };
+use contexts::merge_readiness::domain::cache::{CachePort, CacheState};
 use contexts::merge_readiness::infrastructure::{gh::GhClient, logger::Logger};
 use contexts::merge_readiness::interface::{
     cli::prompt::{self, AFTER_HELP, PromptArgs},
-    presentation::PresentationConfigPort,
+    presentation::{Presenter, PresentationConfigPort},
 };
+use contexts::status_cache::application::cache::{self as status_cache_app, CacheQueryResult};
+use contexts::status_cache::infrastructure::daemon_client::DaemonClient;
 
 #[derive(Parser)]
 #[command(
@@ -106,6 +107,20 @@ impl PresentationConfigPort for ConfigAdapter {
     }
 }
 
+/// daemon 経由でキャッシュを参照する `CachePort` アダプタ（bin 層）
+struct DaemonCacheAdapter;
+
+impl CachePort for DaemonCacheAdapter {
+    fn check(&self, repo_id: &str) -> CacheState {
+        match status_cache_app::query(&DaemonClient, repo_id) {
+            CacheQueryResult::Fresh(s) => CacheState::Fresh(s),
+            CacheQueryResult::Stale(s) => CacheState::Stale(s),
+            // Miss: daemon が refresh 予約済み。Unavailable: lazy_start 済み → next call でヒット
+            CacheQueryResult::Miss | CacheQueryResult::Unavailable => CacheState::Miss,
+        }
+    }
+}
+
 fn main() {
     let repo_id_port = InfraRepoIdPort;
     match Cli::parse().command {
@@ -118,7 +133,17 @@ fn main() {
                         ConfigAdapter::load(),
                     );
                 }
-                ExecutionMode::Cached => cached::run(),
+                ExecutionMode::Cached => {
+                    let cache = DaemonCacheAdapter;
+                    match contexts::merge_readiness::application::prompt::resolve_cached(
+                        &repo_id_port,
+                        &cache,
+                    ) {
+                        PromptEffect::NoOutput => {}
+                        PromptEffect::Show(s) | PromptEffect::ShowAndRefresh(s) => print!("{s}"),
+                        PromptEffect::ShowLoadingAndRefresh => print!("? loading"),
+                    }
+                }
             }
         }
         Some(Command::Config { subcommand }) => match subcommand {
@@ -158,7 +183,17 @@ fn main() {
                 DaemonCommand::Status => {
                     contexts::status_cache::interface::cli::daemon::status(&lifecycle);
                 }
-                DaemonCommand::Refresh { repo_id } => refresh::run(&repo_id),
+                DaemonCommand::Refresh { repo_id } => {
+                    let tokens = contexts::merge_readiness::application::prompt::fetch_output(
+                        &GhClient::new(),
+                        &Logger,
+                    );
+                    if let Some(tokens) = tokens {
+                        let output =
+                            Presenter::new(ConfigAdapter::load()).render_to_string(&tokens);
+                        status_cache_app::update(&DaemonClient, &repo_id, &output);
+                    }
+                }
             }
         }
         None => {
